+++ lib/protobuf/cli.rb
@@ -5,6 +5,7 @@ require 'protobuf/version'
 require 'protobuf/logging'
 require 'protobuf/rpc/servers/socket_runner'
 require 'protobuf/rpc/servers/zmq_runner'
+require 'protobuf/rpc/servers/http_runner'
 
 module Protobuf
   class CLI < ::Thor
@@ -33,6 +34,7 @@ module Protobuf
 
     option :socket,                     :type => :boolean, :aliases => %w(-s), :desc => 'Socket Mode for server and client connections.'
     option :zmq,                        :type => :boolean, :aliases => %w(-z), :desc => 'ZeroMQ Socket Mode for server and client connections.'
+    option :http,                       :type => :boolean, :aliases => %w(), :desc => 'HTTP Server Mode for server and client connections.'
 
     option :beacon_interval,            :type => :numeric, :desc => 'Broadcast beacons every N seconds. (default: 5)'
     option :beacon_port,                :type => :numeric, :desc => 'Broadcast beacons to this port (default: value of ServiceDirectory.port)'
@@ -120,6 +122,8 @@ module Protobuf
                       :socket
                     elsif options.zmq?
                       :zmq
+                    elsif options.http?
+                      :http
                     else
                       case server_type = ENV["PB_SERVER_TYPE"]
                       when nil, /socket/i
@@ -159,6 +163,8 @@ module Protobuf
                         create_zmq_runner
                       when :socket
                         create_socket_runner
+                      when :http
+                        create_http_runner
                       else
                         say_and_exit("Unknown runner mode: #{mode}")
                       end
@@ -218,10 +224,17 @@ module Protobuf
         self.runner = ::Protobuf::Rpc::ZmqRunner.new(runner_options)
       end
 
+      def create_http_runner
+        require 'protobuf/http'
+
+        self.runner = ::Protobuf::Rpc::HttpRunner.new(runner_options)
+      end
+
       def shutdown_server
         logger.info { 'RPC Server shutting down...' }
         runner.stop
         ::Protobuf::Rpc::ServiceDirectory.instance.stop
+        logger.info { 'Shutdown complete' }
       end
 
       # Start the runner and log the relevant options.