+++ lib/protobuf/rpc/error/server_error.rb
@@ -39,5 +39,23 @@ module Protobuf
       end
     end
 
+    class UnauthorizedRequest < PbError
+      def initialize message='The request requires user authentication'
+        super message, 'UNAUTHORIZED_REQUEST'
+      end
+    end
+
+    class ForbiddenRequest < PbError
+      def initialize message='User authentificated but does not have permissions'
+        super message, 'FORBIDDEN_REQUEST'
+      end
+    end
+
+    class DataNotFound < PbError
+      def initialize message='Requested data not found'
+        super message, 'DATA_NOT_FOUND'
+      end
+    end
+
   end
 end